readme updates & setup/workers move
